+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 (Add entries below in the order of server, console, cli, docs, others)
 
+
+- server: fix v2 -> v1 downgrade bug when cron triggers exist
 - server: add index on the `event_id` column of the `hdb_cron_event_invocation_logs` table
 - server: fix GraphQL type for remote relationship field (close #7284)
 - server: support EdDSA algorithm and key type for JWT